+++ crates/ryft-core/src/types/data_type.rs
@@ -1,13 +1,44 @@
+use std::backtrace::Backtrace;
 use std::fmt::Display;
 
 use ryft_macros::Parameter;
+use thiserror::Error;
 
 #[cfg(feature = "xla")]
 use ryft_pjrt::BufferType;
 
-use crate::errors::Error;
 use crate::parameters::Parameter;
 use crate::types::Type;
+
+/// Represents [`DataType`]-related errors.
+#[derive(Error, Clone, Debug, Eq, PartialEq, Hash)]
+pub enum DataTypeError {
+    #[error("{message}")]
+    InvalidDataType { message: String, backtrace: String },
+
+    #[error("cannot construct a promoted data type from an empty collection of data types")]
+    EmptyDataTypePromotionInput { backtrace: String },
+
+    #[error("{message}")]
+    InvalidDataTypePromotion { message: String, backtrace: String },
+}
+
+impl DataTypeError {
+    /// Creates a new [`DataTypeError::InvalidDataType`].
+    pub fn invalid_data_type<M: Into<String>>(message: M) -> Self {
+        Self::InvalidDataType { message: message.into(), backtrace: Backtrace::capture().to_string() }
+    }
+
+    /// Creates a new [`DataTypeError::EmptyDataTypePromotionInput`].
+    pub fn empty_data_type_promotion_input() -> Self {
+        Self::EmptyDataTypePromotionInput { backtrace: Backtrace::capture().to_string() }
+    }
+
+    /// Creates a new [`DataTypeError::InvalidDataTypePromotion`].
+    pub fn invalid_data_type_promotion<M: Into<String>>(message: M) -> Self {
+        Self::InvalidDataTypePromotion { message: message.into(), backtrace: Backtrace::capture().to_string() }
+    }
+}
 
 #[cfg_attr(doc, aquamarine::aquamarine)]
 /// Represents a primitive data type that can be stored in arrays, tensors, matrices, vectors, scalars, etc., which
@@ -1278,8 +1309,9 @@ impl DataType {
     /// This operation is order-invariant, meaning that it returns the same result irrespective of the order in which
     /// the input data types are provided.
     ///
-    /// This function returns [`Error::InvalidArgument`] when `data_types` is empty and
-    /// [`Error::InvalidDataTypePromotion`] when the provided types are incompatible in terms of type promotion.
+    /// This function returns [`DataTypeError::EmptyDataTypePromotionInput`] when `data_types` is empty and
+    /// [`DataTypeError::InvalidDataTypePromotion`] when the provided types are incompatible in terms of type
+    /// promotion.
     ///
     /// # Examples
     ///
@@ -1293,28 +1325,28 @@ impl DataType {
     /// assert_eq!(DataType::promoted(&[DataType::I64, DataType::U64]), Ok(DataType::F64));
     /// ```
     #[inline]
-    pub fn promoted(data_types: &[Self]) -> Result<Self, Error> {
+    pub fn promoted(data_types: &[Self]) -> Result<Self, DataTypeError> {
         let Some((head, tail)) = data_types.split_first() else {
-            return Err(Error::invalid_argument(
-                "cannot construct a promoted data type from an empty collection of data types",
-            ));
+            return Err(DataTypeError::empty_data_type_promotion_input());
         };
         tail.iter().try_fold(*head, |x, y| {
             DATA_TYPE_PROMOTION_LEAST_UPPER_BOUNDS[x.index()][y.index()].ok_or_else(|| {
-                Error::invalid_data_type_promotion(format!("cannot promote types `{x}` and `{y}` to a common type"))
+                DataTypeError::invalid_data_type_promotion(format!(
+                    "cannot promote types `{x}` and `{y}` to a common type",
+                ))
             })
         })
     }
 
     /// Promotes this [`DataType`] to the provided [`DataType`]. This function returns
-    /// [`Error::InvalidDataTypePromotion`] when the promotion is not allowed. Refer to the documentation
-    /// of [`DataType`] for more information on type promotions and the rules that govern them.
+    /// [`DataTypeError::InvalidDataTypePromotion`] when the promotion is not allowed. Refer to the documentation of
+    /// [`DataType`] for more information on type promotions and the rules that govern them.
     #[inline]
-    pub fn promote_to(self, other: DataType) -> Result<DataType, Error> {
+    pub fn promote_to(self, other: DataType) -> Result<DataType, DataTypeError> {
         if self.is_promotable_to(other) {
             Ok(other)
         } else {
-            Err(Error::invalid_data_type_promotion(format!("cannot promote type `{self}` to type `{other}`")))
+            Err(DataTypeError::invalid_data_type_promotion(format!("cannot promote type `{self}` to type `{other}`")))
         }
     }
 
@@ -1386,9 +1418,9 @@ impl Type for DataType {
 
 #[cfg(feature = "xla")]
 impl DataType {
-    /// Creates a [`DataType`] from the provided PJRT [`BufferType`]. Returns [`Error::InvalidDataType`] when the
-    /// provided [`BufferType`] is [`BufferType::Invalid``], which is a PJRT-internal sentinel value.
-    pub fn from_pjrt_buffer_type(buffer_type: BufferType) -> Result<Self, Error> {
+    /// Creates a [`DataType`] from the provided PJRT [`BufferType`]. Returns [`DataTypeError::InvalidDataType`]
+    /// when the provided [`BufferType`] is [`BufferType::Invalid``], which is a PJRT-internal sentinel value.
+    pub fn from_pjrt_buffer_type(buffer_type: BufferType) -> Result<Self, DataTypeError> {
         buffer_type.try_into()
     }
 
@@ -1400,12 +1432,12 @@ impl DataType {
 
 #[cfg(feature = "xla")]
 impl TryFrom<BufferType> for DataType {
-    type Error = Error;
+    type Error = DataTypeError;
 
     fn try_from(buffer_type: BufferType) -> Result<Self, Self::Error> {
         match buffer_type {
             BufferType::Invalid => {
-                Err(Error::invalid_data_type(format!("invalid data type from PJRT: '{buffer_type}'")))
+                Err(DataTypeError::invalid_data_type(format!("invalid data type from PJRT: '{buffer_type}'")))
             }
             BufferType::Token => Ok(Self::Token),
             BufferType::Predicate => Ok(Self::Boolean),
@@ -1483,11 +1515,9 @@ impl From<DataType> for BufferType {
 
 #[cfg(test)]
 mod tests {
-    use crate::errors::Error;
-
     #[cfg(feature = "xla")]
     use super::BufferType;
-    use super::DataType;
+    use super::{DataType, DataTypeError};
 
     #[test]
     fn test_data_type_promoted() {
@@ -1535,44 +1565,40 @@ mod tests {
         assert_eq!(DataType::promoted(&[DataType::F16, DataType::BF16, DataType::F64]), Ok(DataType::F64));
         assert_eq!(DataType::promoted(&[DataType::F64, DataType::F16, DataType::BF16]), Ok(DataType::F64));
         assert_eq!(DataType::promoted(&[DataType::BF16, DataType::F64, DataType::F16]), Ok(DataType::F64));
-        assert!(matches!(
-            DataType::promoted(&[]),
-            Err(Error::InvalidArgument { message, .. })
-                if message == "cannot construct a promoted data type from an empty collection of data types",
-        ));
+        assert!(matches!(DataType::promoted(&[]), Err(DataTypeError::EmptyDataTypePromotionInput { .. }),));
         assert!(matches!(
             DataType::promoted(&[DataType::F8E3M4, DataType::F32]),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote types `f8e3m4` and `f32` to a common type",
         ));
         assert!(matches!(
             DataType::promoted(&[DataType::F32, DataType::F8E3M4]),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote types `f32` and `f8e3m4` to a common type",
         ));
         assert!(matches!(
             DataType::promoted(&[DataType::U4, DataType::F8E3M4]),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote types `u4` and `f8e3m4` to a common type",
         ));
         assert!(matches!(
             DataType::promoted(&[DataType::Boolean, DataType::U1, DataType::F8E3M4]),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote types `u1` and `f8e3m4` to a common type",
         ));
         assert!(matches!(
             DataType::promoted(&[DataType::F8E3M4, DataType::F8E8M0FNU]),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote types `f8e3m4` and `f8e8m0fnu` to a common type",
         ));
         assert!(matches!(
             DataType::promoted(&[DataType::F8E8M0FNU, DataType::F8E3M4]),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote types `f8e8m0fnu` and `f8e3m4` to a common type",
         ));
         assert!(matches!(
             DataType::promoted(&[DataType::F4E2M1FN, DataType::F8E3M4]),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote types `f4e2m1fn` and `f8e3m4` to a common type",
         ));
     }
@@ -1593,36 +1619,37 @@ mod tests {
         assert_eq!(DataType::F64.promote_to(DataType::C128), Ok(DataType::C128));
         assert!(matches!(
             DataType::Boolean.promote_to(DataType::I1),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote type `bool` to type `i1`",
         ));
         assert!(matches!(
             DataType::U4.promote_to(DataType::I4),
-            Err(Error::InvalidDataTypePromotion { message, .. }) if message == "cannot promote type `u4` to type `i4`",
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
+                if message == "cannot promote type `u4` to type `i4`",
         ));
         assert!(matches!(
             DataType::I16.promote_to(DataType::Boolean),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote type `i16` to type `bool`",
         ));
         assert!(matches!(
             DataType::F8E5M2FNUZ.promote_to(DataType::BF16),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote type `f8e5m2fnuz` to type `bf16`",
         ));
         assert!(matches!(
             DataType::F8E3M4.promote_to(DataType::F4E2M1FN),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote type `f8e3m4` to type `f4e2m1fn`",
         ));
         assert!(matches!(
             DataType::F8E3M4.promote_to(DataType::F8E4M3FN),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote type `f8e3m4` to type `f8e4m3fn`",
         ));
         assert!(matches!(
             DataType::F64.promote_to(DataType::C64),
-            Err(Error::InvalidDataTypePromotion { message, .. })
+            Err(DataTypeError::InvalidDataTypePromotion { message, .. })
                 if message == "cannot promote type `f64` to type `c64`",
         ));
     }
@@ -1672,7 +1699,7 @@ mod tests {
     fn test_data_type_from_and_to_pjrt_buffer_type() {
         assert!(matches!(
             DataType::from_pjrt_buffer_type(BufferType::Invalid),
-            Err(Error::InvalidDataType { message, .. }) if message == "invalid data type from PJRT: 'invalid'",
+            Err(DataTypeError::InvalidDataType { message, .. }) if message == "invalid data type from PJRT: 'invalid'",
         ));
         for &(data_type, buffer_type) in &[
             (DataType::Token, BufferType::Token),