+++ crates/ryft-core/src/types_v0/array_structure_type.rs
@@ -142,7 +142,7 @@ pub enum ArrayStructureTypeBroadcastingError {
 
 #[cfg(test)]
 mod tests {
-    use crate::types::DataType::*;
+    use crate::types::data_type::DataType::*;
     use crate::types_v0::array_type::{Shape, Size};
 
     use super::*;