+++ crates/ryft-core/src/types_v0/array_type.rs
@@ -10,9 +10,8 @@ use thiserror::Error;
 
 use ryft_macros::Parameter;
 
-use crate::errors::Error as CoreError;
 use crate::parameters::Parameter;
-use crate::types::DataType;
+use crate::types::data_type::{DataType, DataTypeError as CoreDataTypeError};
 use crate::types_v0::r#type::Type;
 
 impl Type for DataType {
@@ -299,7 +298,8 @@ pub enum ShapeBroadcastingError {
 /// # Examples
 ///
 /// ```rust
-/// # use ryft_core::types_v0::array_type::{ArrayType, DataType, Shape, Size};
+/// # use ryft_core::types::DataType;
+/// # use ryft_core::types_v0::array_type::{ArrayType, Shape, Size};
 ///
 /// // Boolean scalar.
 /// assert_eq!(
@@ -353,7 +353,8 @@ impl ArrayType {
     /// # Examples
     ///
     /// ```rust
-    /// # use ryft_core::types_v0::array_type::{ArrayType, DataType, Shape, Size};
+    /// # use ryft_core::types::DataType;
+    /// # use ryft_core::types_v0::array_type::{ArrayType, Shape, Size};
     ///
     /// // Boolean scalar.
     /// assert_eq!(ArrayType::new(DataType::Boolean, Shape::scalar()).rank(), 0);
@@ -394,7 +395,8 @@ impl ArrayType {
     /// # Examples
     ///
     /// ```rust
-    /// # use ryft_core::types_v0::array_type::{ArrayType, DataType, Shape, Size};
+    /// # use ryft_core::types::DataType;
+    /// # use ryft_core::types_v0::array_type::{ArrayType, Shape, Size};
     /// let w = ArrayType::new(DataType::F32, Shape::new(vec![Size::Static(42), Size::Static(42)]));
     /// let x = ArrayType::new(DataType::Boolean, Shape::scalar());
     /// let y = ArrayType::new(DataType::U16, Shape::new(vec![Size::Dynamic(Some(10))]));
@@ -438,7 +440,8 @@ impl ArrayType {
     /// # Examples
     ///
     /// ```rust
-    /// # use ryft_core::types_v0::array_type::{ArrayType, DataType, Shape, Size};
+    /// # use ryft_core::types::DataType;
+    /// # use ryft_core::types_v0::array_type::{ArrayType, Shape, Size};
     /// let w = ArrayType::new(DataType::F32, Shape::new(vec![Size::Static(42), Size::Static(42)]));
     /// let x = ArrayType::new(DataType::Boolean, Shape::scalar());
     /// let y = ArrayType::new(DataType::U16, Shape::new(vec![Size::Dynamic(Some(10))]));
@@ -497,7 +500,7 @@ pub enum ArrayTypeBroadcastingError {
     /// Error returned when failing to broadcast two [`ArrayType`]s due to their [`DataType`]s not being
     /// compatible (i.e., the [`DataType`] of one cannot be promoted to the [`DataType`] of the other).
     #[error("{0}")]
-    IncompatibleDataTypes(#[from] CoreError),
+    IncompatibleDataTypes(#[from] CoreDataTypeError),
 
     /// Error returned when failing to broadcast two [`ArrayType`]s due to their [`Shape`]s not being
     /// compatible (i.e., the [`Shape`] of one cannot be broadcast to the [`Shape`] of the other).